Replace reg dep RealRegister property checks with meaningful queries
Register dependencies rely on "special" RealRegister values to convey
other properties about the register dependency, such as whether a byte
register must be assigned, or whether any assignable register is
suitable for this dependency.  The long-term goal is to not rely on
such special values in the RealRegister enum to convey these properties.
To get there, this PR replaces direct tests of the real register for
queries on the register dependency itself for the property in question.

This PR also fixes a number of compiler warnings on Z where
`RealRegister::RegNum` and `RealRegister::RegDep` enums were compared
directly to test for these properties.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>