Correct trace messages for LiveVariableInformation
A set of trace messages in the TR_LiveVariableInformation method,
visitTreeForLocals, report during a backwards tree walk whether a node
contains the first commoned reference for a load of a sym.  The way the
messages are currently reported, they might indicate that a node that
does not have a symbol reference contains the first commoned reference
to a sym.  Corrected this to ensure the message only mentions syms for
node that are known to contain references to local symbols.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>