Inline short, primitive, forward arraycopies on AArch64
Forward arraycopies are expected to be the common case, and investigation on
other platforms suggests that arraycopies <= 63 bytes are important in a number
of workloads and will benefit from being specialized inline.

The feature is enabled under `AGGRESSIVE_THROUGHPUT` in OMROptions.cpp.

This optimization can be disabled by setting the environment variable
`TR_DisableInlinePrimitiveForwardArraycopy`

Signed-off-by: Daryl Maier <maier@ca.ibm.com>