Create a log filename builder
* Consolidate and simplify the construction of log file names by creating a
builder that synthesizes a file name from a base name, an optional numerical
suffix, and an optional suffix that can contain special format specifiers.  The
suffix is processed by the port library, if one exists.  The format specifiers
are those described in the documentation for 'omrstr_create_tokens()' in
port/common/omrstr.c.

* Remove now-obsolete FrontEnd name formatter and non-port library code for
collecting the pid and current time

* Rename some identifiers to match what their actual semantics are

* Rename enable/disable compiler options for applying log file name suffixes to
be more descriptive

Signed-off-by: Daryl Maier <maier@ca.ibm.com>